Features/624 skills

+++ sources/JobMagnet/Infrastructure/Entities/SkillItemEntity.cs
@@ -7,9 +7,9 @@ namespace JobMagnet.Infrastructure.Entities;
 public class SkillItemEntity : TrackableEntity<long>
 {
     public string Name { get; set; } // Nombre de la habilidad (ej. "C#", "React")
-    public int ProficiencyLevel { get; set; } // Nivel de habilidad (1-10)
+    public ushort ProficiencyLevel { get; set; } // Nivel de habilidad (1-10)
     public string Category { get; set; } // Categoría (ej. "Programación", "Diseño")
-    public int Rank { get; set; }
+    public ushort Rank { get; set; }
     public string IconUrl { get; set; }
 
     [ForeignKey(nameof(Skill))] public long SkillId { get; set; }
